+++ source/main/cardano/CardanoWalletLauncher.js
@@ -1,64 +1,119 @@
 // @flow
-import { spawn } from 'child_process';
-import { dirname } from 'path';
-import type { ChildProcess } from 'child_process';
+import { merge } from 'lodash';
+import * as cardanoLauncher from 'cardano-launcher';
+import type { Launcher } from 'cardano-launcher';
+import type { NodeConfig } from '../config';
 import { STAKE_POOL_REGISTRY_URL } from '../config';
-import { environment } from '../environment';
-import { NIGHTLY, SELFNODE, QA } from '../../common/types/environment.types';
+import {
+  NIGHTLY,
+  SELFNODE,
+  QA,
+  ITN_REWARDS_V1,
+} from '../../common/types/environment.types';
 import { Logger } from '../utils/logging';
 
 export type WalletOpts = {
-  path: string,
-  walletArgs: string[],
-  nodeBin: string,
   nodeImplementation: 'jormungandr' | 'cardano-node',
-  logStream: any,
+  nodeConfig: NodeConfig,
   cluster: string,
+  stateDir: string,
+  block0Path: string,
+  block0Hash: string,
+  secretPath: string,
+  configPath: string,
+  syncTolerance: string,
+  logFile: any,
 };
 
-export async function CardanoWalletLauncher(
-  walletOpts: WalletOpts
-): Promise<ChildProcess> {
+export function CardanoWalletLauncher(walletOpts: WalletOpts): Launcher {
   const {
-    logStream,
     nodeImplementation,
-    nodeBin,
-    path,
-    walletArgs,
+    nodeConfig, // For cardano-node / byron only!
     cluster,
+    stateDir,
+    block0Path,
+    block0Hash,
+    secretPath,
+    configPath,
+    syncTolerance,
+    logFile,
   } = walletOpts;
-  const walletStdio: string[] = ['pipe', 'pipe', 'pipe', 'ipc'];
-  const nodePath = dirname(nodeBin);
-  const PATH: string = (process.env.PATH: any);
-  const envVariables: $Exact<{
-    PATH: string,
-    CARDANO_WALLET_STAKE_POOL_REGISTRY_URL?: string,
-  }> = {
-    PATH: `${nodePath}:${PATH}`,
+  // TODO: Update launcher config to pass number
+  const syncToleranceSeconds = parseInt(syncTolerance.replace('s', ''), 10);
+
+  // Shared launcher config (node implementations agnostic)
+  const launcherConfig = {
+    networkName: cluster,
+    stateDir,
+    nodeConfig: {
+      kind: nodeImplementation,
+      configurationDir: '',
+      network: {
+        configFile: configPath,
+      },
+    },
+    syncToleranceSeconds,
+    childProcessLogWriteStream: logFile,
   };
 
   // This switch statement handles any node specifc
   // configuration, prior to spawning the child process
   Logger.info('Node implementation', { nodeImplementation });
   switch (nodeImplementation) {
-    case 'cardano-node':
+    case 'cardano':
+      merge(launcherConfig, { nodeConfig });
       break;
     case 'jormungandr':
-      if (cluster === 'selfnode') {
-        Object.assign(envVariables, {
-          CARDANO_WALLET_STAKE_POOL_REGISTRY_URL:
-            STAKE_POOL_REGISTRY_URL[SELFNODE],
+      if (cluster === SELFNODE) {
+        merge(launcherConfig, {
+          apiPort: 8088,
+          nodeConfig: {
+            restPort: 8888,
+            network: {
+              genesisBlock: {
+                file: block0Path,
+                hash: block0Hash,
+              },
+              secretFile: [secretPath],
+            },
+          },
+          stakePoolRegistryUrl: STAKE_POOL_REGISTRY_URL[SELFNODE],
         });
       }
-      if (environment.isIncentivizedTestnetNightly) {
-        Object.assign(envVariables, {
-          CARDANO_WALLET_STAKE_POOL_REGISTRY_URL:
-            STAKE_POOL_REGISTRY_URL[NIGHTLY],
+      if (cluster === NIGHTLY) {
+        merge(launcherConfig, {
+          nodeConfig: {
+            network: {
+              genesisBlock: {
+                hash: block0Hash,
+              },
+            },
+          },
+          stakePoolRegistryUrl: STAKE_POOL_REGISTRY_URL[NIGHTLY],
         });
       }
-      if (environment.isIncentivizedTestnetQA) {
-        Object.assign(envVariables, {
-          CARDANO_WALLET_STAKE_POOL_REGISTRY_URL: STAKE_POOL_REGISTRY_URL[QA],
+      if (cluster === QA) {
+        merge(launcherConfig, {
+          nodeConfig: {
+            network: {
+              genesisBlock: {
+                hash: block0Hash,
+              },
+            },
+          },
+          stakePoolRegistryUrl: STAKE_POOL_REGISTRY_URL[QA],
+        });
+      }
+      if (cluster === ITN_REWARDS_V1) {
+        merge(launcherConfig, {
+          nodeConfig: {
+            network: {
+              genesisBlock: {
+                file: block0Path,
+                hash: block0Hash,
+              },
+            },
+          },
         });
       }
       break;
@@ -66,18 +121,10 @@ export async function CardanoWalletLauncher(
       break;
   }
 
-  Logger.info('Starting Node now...', { path, walletArgs });
-  const childProcess = spawn(path, walletArgs, {
-    stdio: walletStdio,
-    env: {
-      // $FlowFixMe
-      ...process.env,
-      ...envVariables,
-    },
+  Logger.info('Setting up CardanoLauncher now...', {
+    walletOpts,
+    launcherConfig,
   });
 
-  childProcess.stdout.on('data', data => logStream.write(data));
-  childProcess.stderr.on('data', data => logStream.write(data));
-
-  return childProcess;
+  return new cardanoLauncher.Launcher(launcherConfig);
 }